feat(gpud): separate command for "gpud scan"
